Refactor Loki so LokiQueryField can be used independently
