+++ public/app/plugins/datasource/loki/components/AnnotationsQueryEditor.tsx
@@ -1,9 +1,9 @@
 // Libraries
 import React, { memo } from 'react';
-
 // Types
 import { LokiQuery } from '../types';
-import { LokiQueryFieldForm } from './LokiQueryFieldForm';
+import { LokiQueryField } from './LokiQueryField';
+import { LokiOptionFields } from './LokiOptionFields';
 import LokiDatasource from '../datasource';
 
 interface Props {
@@ -18,7 +18,7 @@ export const LokiAnnotationsQueryEditor = memo(function LokiAnnotationQueryEdito
   const { expr, maxLines, instant, datasource, onChange } = props;
 
   // Timerange to get existing labels from. Hard coding like this seems to be good enough right now.
-  const absolute = {
+  const absoluteRange = {
     from: Date.now() - 10000,
     to: Date.now(),
   };
@@ -31,13 +31,23 @@ export const LokiAnnotationsQueryEditor = memo(function LokiAnnotationQueryEdito
   };
   return (
     <div className="gf-form-group">
-      <LokiQueryFieldForm
+      <LokiQueryField
         datasource={datasource}
         query={queryWithRefId}
         onChange={onChange}
         onRunQuery={() => {}}
+        onBlur={() => {}}
         history={[]}
-        absoluteRange={absolute}
+        absoluteRange={absoluteRange}
+        ExtraFieldElement={
+          <LokiOptionFields
+            queryType={queryWithRefId.instant ? 'instant' : 'range'}
+            lineLimitValue={queryWithRefId?.maxLines?.toString() || ''}
+            query={queryWithRefId}
+            onRunQuery={() => {}}
+            onChange={onChange}
+          />
+        }
       />
     </div>
   );
