+++ public/app/plugins/datasource/loki/components/LokiExploreQueryEditor.tsx
@@ -6,11 +6,13 @@ import { ExploreQueryFieldProps } from '@grafana/data';
 import { LokiDatasource } from '../datasource';
 import { LokiQuery, LokiOptions } from '../types';
 import { LokiQueryField } from './LokiQueryField';
+import { LokiOptionFields } from './LokiOptionFields';
 
 type Props = ExploreQueryFieldProps<LokiDatasource, LokiQuery, LokiOptions>;
 
 export function LokiExploreQueryEditor(props: Props) {
   const { range, query, data, datasource, history, onChange, onRunQuery } = props;
+  const absoluteTimeRange = { from: range!.from!.valueOf(), to: range!.to!.valueOf() }; // Range here is never optional
 
   return (
     <LokiQueryField
@@ -21,7 +23,16 @@ export function LokiExploreQueryEditor(props: Props) {
       onRunQuery={onRunQuery}
       history={history}
       data={data}
-      range={range}
+      absoluteRange={absoluteTimeRange}
+      ExtraFieldElement={
+        <LokiOptionFields
+          queryType={query.instant ? 'instant' : 'range'}
+          lineLimitValue={query?.maxLines?.toString() || ''}
+          query={query}
+          onRunQuery={onRunQuery}
+          onChange={onChange}
+        />
+      }
     />
   );
 }
